consolidated user/admin views and navbar to single view and navbar
