+++ src/components/nav/NavBar.js
@@ -1,19 +1,42 @@
-//  updated for Applyances
+import { Link, useNavigate } from "react-router-dom"
 import "./NavBar.css"
-import { AdminNavBar } from "./AdminNavBar"
-import { UserNavBar } from "./UserNavBar"
+import { ApplyanceSearch } from "../applyances/ApplyanceSearch"
 
 export const NavBar = () => {
+    const navigate = useNavigate()
 
 
-    // get apply user object out of local storage
-    const localApplyUser = localStorage.getItem("apply_user") // a string
-    const applyUserObject = JSON.parse(localApplyUser) // an object with 2 keys (id and staff)
 
-    if (applyUserObject.admin) {
-        return <AdminNavBar />
-    }
-    else {
-        return <UserNavBar />
-    }
+    return (
+        <ul className="navbar">
+            <li className="navbar__item navbar__home">
+                <Link className="navbar__link" to="/home">Home</Link>
+            </li>
+
+            <li className="navbar__item navbar__all">
+                <Link className="navbar__link" to="/all-applyances">All ApplYances</Link>
+            </li>
+
+            <li className="navbar__item navbar__my">
+                <Link className="navbar__link" to="/my-applyances">My ApplYances</Link>
+            </li>
+
+            {/* <li className="navbar__item navbar__search">
+                <ApplyanceSearch /> 
+                </li> */}
+
+            <li className="navbar__item navbar__add">
+                <Link className="navbar__link" to="/new-applyance">Add New ApplYance</Link>
+            </li>
+
+
+
+            <li className="navbar__item navbar__logout">
+                <Link className="navbar__link" to="" onClick={() => {
+                    localStorage.removeItem("apply_user")
+                    navigate("/", { replace: true })
+                }}>Logout</Link>
+            </li>
+        </ul>
+    )
 }
