+++ zLinks.md
@@ -1,21 +1,11 @@
-KitchenAidMixer
-/images/KitchenAidMixer.jpg
-https://drive.google.com/file/d/1x_v2_mxZzU_hdtKJNFtd4rqp0P9jy0ap/view?usp=share_link
-
-
-SharpCarouselConvectionMicrowaveOven
-/images/SharpCarouselConvectionMicrowaveOven.jpg
-https://drive.google.com/file/d/11lCX0LVrtEJE5Zp2EmATyb76BGsP6HFZ/view?usp=share_link
-
-
-MaytagWasher
+Maytag Washer
 /images/MaytagWasher.jpg
 https://drive.google.com/file/d/17zVvObQg5Z1dy9W8cz9UtjzUvQM1wSMY/view?usp=share_link
 
-MaytagDryer
+Maytag Dryer
 /images/MaytagDryer.jpg
 https://drive.google.com/file/d/1fIf5DFV-vGRdmX5geTWHjEsEV3Mh7Uij/view?usp=share_link
 
-GEDishwasher
+GE Dishwasher
 /images/GEDishwasher.jpg
 https://drive.google.com/file/d/1HgtkEOiioMQayHgxUcV2f9rcNuHbz16A/view?usp=share_link
